Create workflow for automatically releases "build"
(previous actions discarded in place of this new workflow)

As part of this action, we check the bundle.yaml is up to date, run linters, unit tests and build the binary.
If the branch is named 'release/v*' we also push new images to both Dockerhub and redhat registries.
Failures on any step stop the flow.

The step for pushing to the redhat registry will be updated shortly with new creds

Update makefile (order and comments)

Fix every issue picked up by the new linters:
- typos
- run gofmt
- other small changes