Overall improvement of log msgs and address a few TODOs
Errors related to building the metrics with unexpected values
are masked with http.handler errors and it's quite difficult
to find the source so I added explicit error msgs on all of
thos cases.

The overall process will no longer return early when there are
no Ondat volumes from the api request. Doing so hides the
collector metrics when that's the case. For consistency on the
returned values I moved those checks into the collectors.