api: added disabledCollectors config option
The validation of config options will be done by the operator's
validation webhook.
Validation tests are going to be part of the operator repo.
The MetricsExporterConfigSpec shall be inlined in
the StorageOSCluster.Spec.Metrics.
Operator will be responsible for ConfigMap's creation and patching based
on the StorageOSCluster.Spec.Metrics values.

Signed-off-by: Michal Minář <michal.minar@ondat.io>