Update resources labels
Remove part of the daemonset that is automatically filled in by kustomize when building the bundle.

Replace the label "category" from the daemonset and subsequently remove it from the service. Also need
to update the public docs example of a ServiceMonitor (https://docs.ondat.io/docs/operations/metric-exporter/)
so the labels match these new ones (when we release a new operator with this change).